refactor: 주차 선택 페이지 재사용 가능하게 변경
