+++ src/pages/admin/WeeklyStatusPage/WeeklyDetailSection/ClosedDetail.tsx
