+++ src/pages/admin/WeeklyStatusPage/WeeklyDetailSection/InProgressDetail.tsx
