+++ src/pages/admin/WeeklyStatusPage/WeeklyDetailSection/OpenDetail.tsx
