+++ src/pages/admin/WeeklyStatusPage/index.tsx
@@ -1,0 +1,8 @@
+import SelectWeekPage from 'pages/SelectWeekPage';
+import React from 'react';
+
+const AdminWeeklyStatusPage = (): JSX.Element => {
+  return <SelectWeekPage isAdmin />;
+};
+
+export default AdminWeeklyStatusPage;
